requirements.txt
pandas
requests
pytest-docker
zmq

